Extracts convertToWebp function into dedicated module
Refactors the `convertToWebp` logic by moving it from `stickerCommand.js` to its own `convertToWebp.js` file.

This improves modularity and separation of concerns, making the sticker processing module more organized and maintainable.
